+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "blitz" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 1.2.0 - 2020-08-14
+
+### Added
+
+- Supported deprecation message.
+
 ## 1.1.0 - 2020-08-06
 
 ### Added